B-Learning-Demo/backend/requirements.txt
# Backend Dependencies for B-Learning Demo
# Python 3.11+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
psycopg2-binary==2.9.9
# Alternative: asyncpg==0.29.0 for async support

# CORS
fastapi-cors==0.0.6

# Environment Variables
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Date/Time
python-dateutil==2.8.2

# Optional: Development tools
pytest==7.4.3
httpx==0.26.0
